backend/requirements.txt
# FastAPI and async server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Pydantic for data validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Database
motor==3.6.0
pymongo==4.9.1

# S3-compatible storage
boto3==1.35.58

# Celery for async tasks
celery[redis]==5.4.0
redis==5.2.0

# PDF processing
pdfplumber==0.11.4
PyPDF2==3.0.1

# OpenAI
openai==1.54.4
tiktoken==0.8.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# HTTP client
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Code quality
flake8==7.1.1
black==24.10.0
isort==5.13.2

# Development
python-dotenv==1.0.1
